requirements.txt
Django
djangorestframework
Pillow
djoser
django-crispy-forms